Fixed visual studio code analysis warning
Warning C6297 Arithmetic overflow: 32-bit value is shifted, then cast to 64-bit
value.  Results might not be an expected value.
c:\compiler\openmsx\src\laserdisc\oggreader.cc line 664

This could indeed theoretically go wrong. I doubt it's also a problem in
practice, but easy enough to fix.

+++ src/laserdisc/OggReader.cc
@@ -429,7 +429,7 @@ size_t OggReader::frameNo(ogg_packet* packet)
 		return size_t(-1);
 	}
 
-	size_t intra = packet->granulepos & ((1 << granuleShift) - 1);
+	size_t intra = packet->granulepos & ((size_t(1) << granuleShift) - 1);
 	size_t key = packet->granulepos >> granuleShift;
 	return key + intra;
 }
@@ -661,7 +661,7 @@ void OggReader::getFrameNo(RawFrame& rawFrame, size_t frameno)
 		}
 
 		// Sanity check, should not happen
-		if (frameList.size() > (2u << granuleShift)) {
+		if (frameList.size() > (size_t(2) << granuleShift)) {
 			// We've got more than twice as many frames
 			// as the maximum distance between key frames.
 			cli.printWarning("Cannot find frame " +
